Rework checkForSiblingStyleChanges to abort earlier
This reorders the operations in checkForSiblingStyleChanges to bail out earlier in
more cases. Specifically if we call setNeedsStyleRecalc() on the element we don't 
need to continue looking for children to call it on. We can also early return if
the element already needs a SubtreeStyleChange or the document has a pending forced
style recalc.

I also swapped out manual versions of firstElementChild, lastElementChild,
nextElementSibling and previousElementSibling with the actual methods and changed it to
not call renderStyle() repeatedly since that contains two branches and a possible
virtual call.

R=eseidel@chromium.org,ojan@chromium.org

Review URL: https://chromiumcodereview.appspot.com/19541002

git-svn-id: svn://svn.chromium.org/blink/trunk@154721 bbb929c8-8fbe-4397-9dbb-9b2b20218538