Move isValid/isCurrentColor from Color to StyleColor
Change style resolution and RenderStyle to use the new StyleColor class
instead of the platform Color class to represent a color and only
convert to the lower level color class at paint/query time.

This allows the currentColor logic to be contained within RenderStyle
and friends, allows style using currentColor to be cached and shared
and hides the complexity from style resolution. It also moves the
isValid concept from Color into StyleColor, thereby reducing the size
of the lower-level Color class to 32 bits.

The SVG code isn't as clean as I'd like it to be, mainly due to the
fact that SVG has a separate currentColor implementation. I'd like to
merge the two or at least make them work together but that'll have to
follow in a separate change as this one is big enough as is.

BUG=259121
R=eseidel@chromium.org,jamesr@chromium.org
TEST=Covered by existing fast/css tests

Review URL: https://chromiumcodereview.appspot.com/20061003

git-svn-id: svn://svn.chromium.org/blink/trunk@154797 bbb929c8-8fbe-4397-9dbb-9b2b20218538