Add a Blink API to set window features on a WebView.
This will allow for re-creating a popup with correct window features
after we've decided to block the popup and not create it previously.

With the old popup blocking behavior, the new WebView was just hidden,
and we never had to set the features manually.

BUG=38458
R=abarth@chromium.org

Review URL: https://chromiumcodereview.appspot.com/19804008

git-svn-id: svn://svn.chromium.org/blink/trunk@154883 bbb929c8-8fbe-4397-9dbb-9b2b20218538