Ensure that settings exists before trying to access it.
In some cases it looks like we try to access the document
settings object before it has been set. Add check (in addition) to
assert to ensure that we do not. This matches the similar checks
we have in the other FontBuilder methods that access Settings.

This is a recent regression from our FontBuilder work.

R=eseidel@chromium.org
BUG=265530

Review URL: https://codereview.chromium.org/21087003

git-svn-id: svn://svn.chromium.org/blink/trunk@155111 bbb929c8-8fbe-4397-9dbb-9b2b20218538