Distribution pool does not need to be a ContentDistribution
ContentDistributor was using a ContentDistribution object for the distribution pool
but this is not needed and we can instead use a Vector<Node*>. This removes lots of
hash table operations when doing distribution since the ContentDistribution was
maintaining a hash table of indexes we never used. We don't even need to keep a RefPtr
to the nodes since they're owned by the document tree during distribution and the pool
never needs to take ownership.

Review URL: https://chromiumcodereview.appspot.com/21090004

git-svn-id: svn://svn.chromium.org/blink/trunk@155121 bbb929c8-8fbe-4397-9dbb-9b2b20218538