Remove ContentDistributor::ensureDistribution
Instead of having this static method we should just do a recursive call up the tree inside
Element::ensureDistribution to update the distribution. This means we don't need to walk the
set of ElementShadows twice and also means that if you're 8+ levels down we don't need to do
any vector allocations. Doing this process recursively should be fine since attach() is
already recursive.

This is the first step in moving toward a recursive top down model for content distribution.

Review URL: https://chromiumcodereview.appspot.com/21123004

git-svn-id: svn://svn.chromium.org/blink/trunk@155127 bbb929c8-8fbe-4397-9dbb-9b2b20218538