Split ScrollbarThemeMac into Overlay and NonOverlay subclasses
This is the first of two steps. The second will move the new
subclasses into their own files.

Move the static variables to be scoped to the classes that access
them (Common, Overlay, or NonOverlay). This separation is fairly
simple, with the exception of gButtonPlacement which (less
obviously) is always ScrollbarButtonsNone for Overlay (and as such
can be made NonOverlay-only.

Duplicate scrollbarMap to scrollbarPainterMap -- this served
two functions before (1) enumerating all scrollbar clients so
that their properties could be updated and (2) mapping overlay
scrollbar clients to their ScrollbarPainter. Now use two separate
structures for the two separate functions.

BUG=133093

Review URL: https://chromiumcodereview.appspot.com/21296003

git-svn-id: svn://svn.chromium.org/blink/trunk@155296 bbb929c8-8fbe-4397-9dbb-9b2b20218538