Update location of w3c shapes tests to reflect new import location
Partial imports should not be imported to the same place as the full
import, so this adds an option to allow importing to a different
location than the default. In doing that, I found that imports from
contributor directories didn't seem to be working correctly, so I
updated the import to be smarter about both contributor directories
and the test status directories. It should now be more likely that the
script guesses the root directory properly.

This is a port from WebKit: http://wkb.ug/118156

BUG=266558

Review URL: https://chromiumcodereview.appspot.com/21406003

git-svn-id: svn://svn.chromium.org/blink/trunk@155334 bbb929c8-8fbe-4397-9dbb-9b2b20218538