[SVG] Position set by low priority animation is ignored even though it is active
In accordance with SMIL animation sandwich model for multiple animations acting on the same element:
http://www.w3.org/TR/2005/REC-SMIL2-20051213/animation.html#animationNS-AnimationSandwichModel

Once the higher priority animation is completed, the animValue should be the value set by low priority animation if it is still active.
In our case, the value is wrongly set to the baseValue even though there is still an active low priority animation.
Modified the logic for higher priority animation to not clear the value set by low priority animation in such a case.

R=pdr@chromium.org, schenney@chromium.org, fmalita@chromium.org, dschulze@chromium.org
BUG=267074

Review URL: https://chromiumcodereview.appspot.com/21624005

git-svn-id: svn://svn.chromium.org/blink/trunk@155414 bbb929c8-8fbe-4397-9dbb-9b2b20218538